+++ src/index.js
@@ -10,7 +10,9 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 export { default as Provider } from "./Provider";
-export { default as Selector, catchDependency } from "./Selector";
+export { default as Selector } from "./Selector";
+export { default as SelectorV3 } from "./SelectorV3";
+export { catchDependency } from "./selectorHelpers";
 export { providers } from "./providers";
 export { storeManager } from "./storeManager";
-export { eventNames } from "./helpers";
+export { eventNames, providerArgsV3 } from "./helpers";