+++ test-e2e/browser/v3-react-app/src/components/delete-icon/DeleteIcon.js
@@ -1,0 +1,16 @@
+import Proptypes from "prop-types";
+
+const DeleteIcon = ({ onClick }) => {
+  return (
+    <span className="delete" onClick={onClick}>
+      {/* eslint-disable-next-line */}
+      🗑️
+    </span>
+  );
+};
+
+DeleteIcon.propTypes = {
+  onClick: Proptypes.func.isRequired,
+};
+
+export default DeleteIcon;