+++ test-e2e/browser/v3-react-app/src/components/placeholder/Placeholder.js
@@ -1,0 +1,5 @@
+const Placeholder = () => {
+  return <li className="placeholder" />;
+};
+
+export default Placeholder;