+++ test-e2e/browser/v3-react-app/src/components/placeholders/Placeholders.js
@@ -1,0 +1,20 @@
+import PropTypes from "prop-types";
+
+import ItemsListContainer from "components/items-list-container";
+import Placeholder from "components/placeholder";
+
+const Placeholders = ({ id }) => {
+  return (
+    <ItemsListContainer id={id} className="placeholders">
+      <Placeholder />
+      <Placeholder />
+      <Placeholder />
+    </ItemsListContainer>
+  );
+};
+
+Placeholders.propTypes = {
+  id: PropTypes.string.isRequired,
+};
+
+export default Placeholders;