+++ test-e2e/browser/v3-react-app/src/data/authors/providers.js
@@ -1,0 +1,25 @@
+import MockProvider from "data/mock-provider";
+
+const AUTHORS = [
+  {
+    id: 1,
+    name: "George Orwell",
+  },
+  {
+    id: 2,
+    name: "Ray Bradbury",
+  },
+  {
+    id: 3,
+    name: "Aldous Huxley",
+  },
+  {
+    id: 4,
+    name: "Ernest Hemingway",
+  },
+];
+
+export const authorsProvider = new MockProvider({
+  id: "authors",
+  data: AUTHORS,
+});