+++ test-e2e/nodejs/v3-cjs/data/authors/index.js
@@ -1,0 +1,10 @@
+const { deleteAuthor, createAuthor } = require("./actions");
+const { authorsProvider } = require("./providers");
+const { authorsSearch } = require("./selectors");
+
+module.exports = {
+  deleteAuthor,
+  createAuthor,
+  authorsProvider,
+  authorsSearch,
+};